Moved data.json to form folder
